requirements.txt
python-dotenv==1.0.0
requests==2.31.0
aiogram==2.25.1
prettytable==3.9.0
